+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - add support for client to pass ack_timeout
 - add is_alive to client api
+- raise special exception on client redis connection error
 
 ## [1.8.9] - 2020-02-05
 ### Fixed